move designers and champions to printing
champion and designer property are now per card
can now support multi-champions and multi-designers on canon reprints
rake no longer complains about every reprint and champion card

there was weirdness going on with CardPrinting::champion and CardPrinting::designer so replaced them with prints_champion and prints_designer.
Card no longer has champion or designer, as they're not needed anymore. an array of @designers and/or @champions could be populated after the @printings array is added to Card if needed in future.